Reacomodo paquetes.
