+++ src/carta/monstruo/ExodiaTheForbiddenOne.java
@@ -1,7 +1,6 @@
 package carta.monstruo;
 
-import carta.Carta;
-import clases.Efecto;
+import carta.Efecto;
 
 public class ExodiaTheForbiddenOne extends CartaMonstruo implements Efecto
 {
