+++ src/carta/monstruo/Jinzo7.java
@@ -1,7 +1,6 @@
 package carta.monstruo;
 
-import carta.Carta;
-import clases.Efecto;
+import carta.Efecto;
 
 public class Jinzo7 extends CartaMonstruo implements Efecto
 {
