+++ src/carta/monstruo/LeftLegOfTheForbiddenOne.java
@@ -1,8 +1,5 @@
 package carta.monstruo;
 
-import carta.Carta;
-import clases.Efecto;
-
 public class LeftLegOfTheForbiddenOne extends CartaMonstruo
 {
