+++ src/carta/monstruo/RightArmOfTheForbiddenOne.java
@@ -1,8 +1,5 @@
 package carta.monstruo;
 
-import carta.Carta;
-import clases.Efecto;
-
 public class RightArmOfTheForbiddenOne extends CartaMonstruo
 {
