+++ src/carta/monstruo/RightLegOfTheForbiddenOne.java
@@ -1,8 +1,5 @@
 package carta.monstruo;
 
-import carta.Carta;
-import clases.Efecto;
-
 public class RightLegOfTheForbiddenOne extends CartaMonstruo
 {
