+++ src/carta/trampa/Reinforcements.java
@@ -1,7 +1,6 @@
 package carta.trampa;
 
-import carta.Carta;
-import clases.Efecto;
+import carta.Efecto;
 
 public class Reinforcements extends CartaTrampa implements Efecto
 {
